Rename identifiers from ElectroluxAppliances to ElectroluxAppliance
