+++ bundles/org.openhab.binding.electroluxappliance/NOTICE
