+++ bundles/org.openhab.binding.electroluxappliance/README.md
