+++ bundles/org.openhab.binding.electroluxappliance/src/main/resources/OH-INF/i18n/electroluxappliance.properties
@@ -1,0 +1,69 @@
+# add-on
+
+addon.electroluxappliance.name = ElectroluxAppliance Binding
+addon.electroluxappliance.description = This is the binding for Electrolux Appliances.
+
+# thing types
+
+thing-type.electroluxappliance.api.label = Electrolux Group API
+thing-type.electroluxappliance.api.description = This bridge represents the API connector.
+thing-type.electroluxappliance.air-purifier.label = Electrolux Pure A9
+thing-type.electroluxappliance.air-purifier.description = This thing represents the Electrolux Air Purifier.
+
+# thing types config
+
+thing-type.config.electroluxappliance.api.apiKey.label = API Key
+thing-type.config.electroluxappliance.api.apiKey.description = Your personal API Key.
+thing-type.config.electroluxappliance.api.refresh.label = Refresh Interval
+thing-type.config.electroluxappliance.api.refresh.description = Specifies the refresh interval in seconds.
+thing-type.config.electroluxappliance.api.accessToken.label = Access Token
+thing-type.config.electroluxappliance.api.accessToken.description = Your personal Access Token.
+thing-type.config.electroluxappliance.api.refreshToken.label = Refresh Token
+thing-type.config.electroluxappliance.api.refreshToken.description = Your personal Refresh Token.
+thing-type.config.electroluxappliance.air-purifier.applianceId.label = Appliance Id
+thing-type.config.electroluxappliance.air-purifier.applianceId.description = Appliance Id.
+
+# channel types
+
+channel-type.electroluxappliance.co2.label = CO2
+channel-type.electroluxappliance.co2.description = The measured CarbonDioxide.
+channel-type.electroluxappliance.door-open.label = Door Status
+channel-type.electroluxappliance.door-open.description = The door status Open/Closed.
+channel-type.electroluxappliance.fan-speed.label = Fan Speed Setting
+channel-type.electroluxappliance.fan-speed.description = The fan speed setting.
+channel-type.electroluxappliance.fan-speed.state.option.1 = Level 1
+channel-type.electroluxappliance.fan-speed.state.option.2 = Level 2
+channel-type.electroluxappliance.fan-speed.state.option.3 = Level 3
+channel-type.electroluxappliance.fan-speed.state.option.4 = Level 4
+channel-type.electroluxappliance.fan-speed.state.option.5 = Level 5
+channel-type.electroluxappliance.fan-speed.state.option.6 = Level 6
+channel-type.electroluxappliance.fan-speed.state.option.7 = Level 7
+channel-type.electroluxappliance.fan-speed.state.option.8 = Level 8
+channel-type.electroluxappliance.fan-speed.state.option.9 = Level 9
+channel-type.electroluxappliance.filter-life.label = Remaining Filter Life
+channel-type.electroluxappliance.filter-life.description = The remaining filter life indication in percent.
+channel-type.electroluxappliance.humidity.label = Humidity
+channel-type.electroluxappliance.humidity.description = The measured humidity.
+channel-type.electroluxappliance.ionizer.label = Ionizer Status
+channel-type.electroluxappliance.ionizer.description = The ionizer status.
+channel-type.electroluxappliance.pm1.label = PM1
+channel-type.electroluxappliance.pm1.description = The Particulate Matter 1 (0.001mm).
+channel-type.electroluxappliance.pm10.label = PM10
+channel-type.electroluxappliance.pm10.description = The Particulate Matter 10 (0.01mm).
+channel-type.electroluxappliance.pm2_5.label = PM2.5
+channel-type.electroluxappliance.pm2_5.description = The Particulate Matter 2.5 (0.0025mm).
+channel-type.electroluxappliance.safety-lock.label = Safety Lock
+channel-type.electroluxappliance.safety-lock.description = the safety lock status.
+channel-type.electroluxappliance.status.label = Current Status
+channel-type.electroluxappliance.status.description = Information on current status.
+channel-type.electroluxappliance.temperature.label = Temperature
+channel-type.electroluxappliance.temperature.description = The measured temperature.
+channel-type.electroluxappliance.tvoc.label = TVOC
+channel-type.electroluxappliance.tvoc.description = The Total Volatile Organic Compounds.
+channel-type.electroluxappliance.ui-light.label = UI Light
+channel-type.electroluxappliance.ui-light.description = The air quality light status indication.
+channel-type.electroluxappliance.work-mode.label = Work Mode Setting
+channel-type.electroluxappliance.work-mode.description = The work mode setting.
+channel-type.electroluxappliance.work-mode.state.option.PowerOff = Power Off
+channel-type.electroluxappliance.work-mode.state.option.Auto = Automatic
+channel-type.electroluxappliance.work-mode.state.option.Manual = Manual
