--- bundles/org.openhab.binding.electroluxappliances/pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.openhab.addons.bundles</groupId>
-    <artifactId>org.openhab.addons.reactor.bundles</artifactId>
-    <version>4.3.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>org.openhab.binding.electroluxappliances</artifactId>
-
-  <name>openHAB Add-ons :: Bundles :: Electrolux Appliances Binding</name>
-
-</project>
