--- bundles/org.openhab.binding.electroluxappliances/src/main/resources/OH-INF/addon/addon.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<addon:addon id="electroluxappliances" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:addon="https://openhab.org/schemas/addon/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/addon/v1.0.0 https://openhab.org/schemas/addon-1.0.0.xsd">
-
-	<type>binding</type>
-	<name>Electrolux Appliances Binding</name>
-	<description>This is the binding for Electrolux Appliances.</description>
-	<connection>cloud</connection>
-
-</addon:addon>
