--- bundles/org.openhab.binding.electroluxappliances/src/main/resources/OH-INF/i18n/electroluxappliance.properties
@@ -1,69 +1,0 @@
-# add-on
-
-addon.electroluxappliances.name = ElectroluxAppliances Binding
-addon.electroluxappliances.description = This is the binding for Electrolux Appliances.
-
-# thing types
-
-thing-type.electroluxappliances.api.label = Electrolux Group API
-thing-type.electroluxappliances.api.description = This bridge represents the API connector.
-thing-type.electroluxappliances.air-purifier.label = Electrolux Pure A9
-thing-type.electroluxappliances.air-purifier.description = This thing represents the Electrolux Air Purifier.
-
-# thing types config
-
-thing-type.config.electroluxappliances.api.apiKey.label = API Key
-thing-type.config.electroluxappliances.api.apiKey.description = Your personal API Key.
-thing-type.config.electroluxappliances.api.refresh.label = Refresh Interval
-thing-type.config.electroluxappliances.api.refresh.description = Specifies the refresh interval in seconds.
-thing-type.config.electroluxappliances.api.accessToken.label = Access Token
-thing-type.config.electroluxappliances.api.accessToken.description = Your personal Access Token.
-thing-type.config.electroluxappliances.api.refreshToken.label = Refresh Token
-thing-type.config.electroluxappliances.api.refreshToken.description = Your personal Refresh Token.
-thing-type.config.electroluxappliances.air-purifier.applianceId.label = Appliance Id
-thing-type.config.electroluxappliances.air-purifier.applianceId.description = Appliance Id.
-
-# channel types
-
-channel-type.electroluxappliances.co2.label = CO2
-channel-type.electroluxappliances.co2.description = The measured CarbonDioxide.
-channel-type.electroluxappliances.door-open.label = Door Status
-channel-type.electroluxappliances.door-open.description = The door status Open/Closed.
-channel-type.electroluxappliances.fan-speed.label = Fan Speed Setting
-channel-type.electroluxappliances.fan-speed.description = The fan speed setting.
-channel-type.electroluxappliances.fan-speed.state.option.1 = Level 1
-channel-type.electroluxappliances.fan-speed.state.option.2 = Level 2
-channel-type.electroluxappliances.fan-speed.state.option.3 = Level 3
-channel-type.electroluxappliances.fan-speed.state.option.4 = Level 4
-channel-type.electroluxappliances.fan-speed.state.option.5 = Level 5
-channel-type.electroluxappliances.fan-speed.state.option.6 = Level 6
-channel-type.electroluxappliances.fan-speed.state.option.7 = Level 7
-channel-type.electroluxappliances.fan-speed.state.option.8 = Level 8
-channel-type.electroluxappliances.fan-speed.state.option.9 = Level 9
-channel-type.electroluxappliances.filter-life.label = Remaining Filter Life
-channel-type.electroluxappliances.filter-life.description = The remaining filter life indication in percent.
-channel-type.electroluxappliances.humidity.label = Humidity
-channel-type.electroluxappliances.humidity.description = The measured humidity.
-channel-type.electroluxappliances.ionizer.label = Ionizer Status
-channel-type.electroluxappliances.ionizer.description = The ionizer status.
-channel-type.electroluxappliances.pm1.label = PM1
-channel-type.electroluxappliances.pm1.description = The Particulate Matter 1 (0.001mm).
-channel-type.electroluxappliances.pm10.label = PM10
-channel-type.electroluxappliances.pm10.description = The Particulate Matter 10 (0.01mm).
-channel-type.electroluxappliances.pm2_5.label = PM2.5
-channel-type.electroluxappliances.pm2_5.description = The Particulate Matter 2.5 (0.0025mm).
-channel-type.electroluxappliances.safety-lock.label = Safety Lock
-channel-type.electroluxappliances.safety-lock.description = the safety lock status.
-channel-type.electroluxappliances.status.label = Current Status
-channel-type.electroluxappliances.status.description = Information on current status.
-channel-type.electroluxappliances.temperature.label = Temperature
-channel-type.electroluxappliances.temperature.description = The measured temperature.
-channel-type.electroluxappliances.tvoc.label = TVOC
-channel-type.electroluxappliances.tvoc.description = The Total Volatile Organic Compounds.
-channel-type.electroluxappliances.ui-light.label = UI Light
-channel-type.electroluxappliances.ui-light.description = The air quality light status indication.
-channel-type.electroluxappliances.work-mode.label = Work Mode Setting
-channel-type.electroluxappliances.work-mode.description = The work mode setting.
-channel-type.electroluxappliances.work-mode.state.option.PowerOff = Power Off
-channel-type.electroluxappliances.work-mode.state.option.Auto = Automatic
-channel-type.electroluxappliances.work-mode.state.option.Manual = Manual
